+++ Mlem/App Constants.swift
@@ -10,7 +10,7 @@ import KeychainAccess
 import UIKit
 
 struct AppConstants {
-    static let cacheSize = 500_000_000 // 500MiB in bytes
+    static let cacheSize = 500_000_000 // 500MB in bytes
     static let urlCache: URLCache = URLCache(memoryCapacity: cacheSize, diskCapacity: cacheSize)
     static let webSocketSession: URLSession = URLSession(configuration: .default)
     static let urlSession: URLSession = URLSession(configuration: .default)
