+++ Mlem/Assets.xcassets/logo.github.symbolset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "symbols" : [
+    {
+      "filename" : "logo.github.svg",
+      "idiom" : "universal"
+    }
+  ]
+}
