+++ Mlem/Views/Tabs/Settings/Components/Views/Appearance/Comment/CommentSettingsView.swift
@@ -8,16 +8,27 @@
 import Foundation
 import SwiftUI
 
-struct CustomizeCommentView: View {
+struct CommentSettingsView: View {
+    
+    @AppStorage("voteComplexOnRight") var shouldShowVoteComplexOnRight: Bool = false
+    @AppStorage("thumbnailsOnRight") var shouldShowThumbnailsOnRight: Bool = false
     // interactions and info
     @AppStorage("commentVoteComplexStyle") var commentVoteComplexStyle: VoteComplexStyle = .standard
     @AppStorage("shouldShowScoreInCommentBar") var shouldShowScoreInCommentBar: Bool = false
     @AppStorage("shouldShowTimeInCommentBar") var shouldShowTimeInCommentBar: Bool = true
     @AppStorage("shouldShowSavedInCommentBar") var shouldShowSavedInCommentBar: Bool = false
     @AppStorage("shouldShowRepliesInCommentBar") var shouldShowRepliesInCommentBar: Bool = true
+    @AppStorage("shouldShowUserServerInComment") var shouldShowUserServerInComment: Bool = false
     
     var body: some View {
         List {
+            Section("Display Sides") {
+                SwitchableSettingsItem(settingPictureSystemName: "arrow.up.arrow.down",
+                                       settingPictureColor: .pink,
+                                       settingName: "Show vote buttons on right",
+                                       isTicked: $shouldShowVoteComplexOnRight)
+            }
+
             Section("Interactions and Info") {
                 SelectableSettingsItem(
                     settingIconSystemName: "arrow.up.arrow.down.square",
@@ -25,6 +36,10 @@ struct CustomizeCommentView: View {
                     currentValue: $commentVoteComplexStyle,
                     options: VoteComplexStyle.allCases
                 )
+                SwitchableSettingsItem(settingPictureSystemName: "server.rack",
+                                                           settingPictureColor: .pink,
+                                                           settingName: "Show user server instance",
+                                                           isTicked: $shouldShowUserServerInComment)
                 SwitchableSettingsItem(settingPictureSystemName: AppConstants.emptyUpvoteSymbolName,
                                        settingPictureColor: .pink,
                                        settingName: "Show upvotes in info",
